fixed ci/cd
