+++ kolezza_dash_board/src/app/(admin)/admin/components/Content/index.tsx
@@ -3,7 +3,7 @@ import React, { useState } from "react";
 import Link from "next/link";
 import { IoIosArrowBack, IoIosArrowForward } from "react-icons/io";
 import { IoMdCheckboxOutline } from "react-icons/io";
-import useGetUsersData from "@/app/hooks/useGetUsersDashboard";
+import useGetUsersData from "@/app/components/hooks/useGetUsersDashboard";
 
 type UserRole = "speech_therapist" | "superadmin";
