+++ kolezza_dash_board/src/app/(admin)/admin/dashboard/page.tsx
@@ -2,11 +2,11 @@
 import React, { useState, useEffect } from 'react';
 import { Chart as ChartJS, ArcElement, Tooltip, Legend, BarElement, CategoryScale, LinearScale} from 'chart.js';
 import { Doughnut, Bar } from 'react-chartjs-2';
-import { useFetchTherapists } from '@/app/hooks/useFetchTherapists';
-import { useFetchChildren } from '@/app/hooks/useFetchChildren';
 import Layout from '../Layout';
 import ChartDataLabels from 'chartjs-plugin-datalabels';
 import { ChartOptions} from 'chart.js';
+import { useFetchTherapists } from '@/app/components/hooks/useFetchTherapists';
+import { useFetchChildren } from '@/app/components/hooks/useFetchChildren';
 
 ChartJS.register(ArcElement, Tooltip, Legend, BarElement, CategoryScale, LinearScale, ChartDataLabels);
