+++ kolezza_dash_board/src/app/Sidebar/index.tsx
@@ -1,53 +1,3 @@
-// import React from 'react';
-// import Image from 'next/image';
-// import Link from 'next/link';
-// import { LayoutDashboard, Users2, UserCircle, Users } from 'lucide-react';
-
-// const Sidebar = () => {
-//   return (
-//     <div className="fixed top-0 left-0 w-64 h-full bg-customDarkBlue text-white flex flex-col rounded-lg shadow-lg">
-//       <div className="p-6 mb-8">
-//         <div className="flex items-center mb-2 mt-7">
-//           <Image 
-//             src="/images/logo.png" 
-//             width={160} 
-//             height={40} 
-//             alt="SawaTok Logo" 
-//             className="mr-2" 
-//           />
-//         </div>
-//       </div>
-//       <nav className="flex-grow -mt-9 nh:w-6 "> 
-//         <ul className="space-y-10 ml-4">
-//           {[
-//             { name: 'Dashboard', icon: LayoutDashboard, href: '/', active: true },
-//             { name: 'Patients', icon: Users2, href: '/patients' },
-//             { name: 'Profile', icon: UserCircle, href: '/profile/ [user_id]' },
-//             { name: 'Users', icon: Users, href:'/users'}, 
-//           ].map((item) => (
-//             <li key={item.name}>
-//               <Link 
-//                 href={item.href} 
-//                 className={`flex items-center px-4 py-3 transition-colors group ${
-//                   item.active ? 'bg-white text-customDarkBlue rounded-2xl w-32' : 'hover:bg-white hover:text-customDarkBlue'
-//                 }`}
-//               >
-//                 <item.icon 
-//                   className={`mr-2 ${item.active ? 'text-green-500' : 'group-hover:text-green-500'}`}
-//                   size={24} 
-//                 />
-//                 <span className="text-lg">{item.name}</span>
-//               </Link>
-//             </li>
-//           ))}
-//         </ul>
-//       </nav>
-//     </div>
-//   );
-// };
-
-// export default Sidebar;
-
 import React, { useState, useEffect } from 'react';
 import Image from 'next/image';
 import Link from 'next/link';
@@ -58,8 +8,8 @@ import { usePathname } from 'next/navigation';
 const Sidebar = () => {
   const [userId, setUserId] = useState<string | null>(null);
   const [isSuperAdmin, setIsSuperAdmin] = useState(false);
-  const [activeLink, setActiveLink] = useState<string>(''); // Track active link
-  const pathname = usePathname(); // Use this for the current pathname
+  const [activeLink, setActiveLink] = useState<string>(''); 
+  const pathname = usePathname(); 
 
   useEffect(() => {
     const id = getCookie('userId') as string | null;
@@ -67,9 +17,8 @@ const Sidebar = () => {
     
     setIsSuperAdmin(pathname.endsWith('admin/users'));
 
-    // Set the initial active link based on the current path
     setActiveLink(pathname);
-  }, [pathname]); // Runs on route change
+  }, [pathname]); 
 
   return (
     <div className="w-64 h-[100%] bg-customDarkBlue text-white flex flex-col">
@@ -94,7 +43,7 @@ const Sidebar = () => {
             <li key={item.name}>
               <Link
                 href={item.href}
-                onClick={() => setActiveLink(item.href)} // Set active link on click
+                onClick={() => setActiveLink(item.href)}
                 className={`flex items-center px-6 py-3 ${
                   activeLink === item.href ? 'bg-white text-customDarkBlue' : 'hover:bg-white hover:text-customDarkBlue'
                 } transition-colors group`}
@@ -109,7 +58,7 @@ const Sidebar = () => {
             <li>
               <Link
                 href="admin/users"
-                onClick={() => setActiveLink('admin/users')} // Set active link on click
+                onClick={() => setActiveLink('admin/users')}
                 className={`flex items-center px-6 py-3 ${
                   activeLink === 'admin/users' ? 'bg-white text-customDarkBlue' : 'hover:bg-white hover:text-customDarkBlue'
                 } transition-colors group`}
