+++ kolezza_dash_board/src/app/components/hooks/useGetChildren.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
-import { fetchChildren } from '@/app/utils/fetchChildren'; // Assuming you have this utility
-import { FetchChildrenResponse, Child } from '@/app/utils/types';
+import { Child, fetchAllChildren } from '@/app/utils/fetchChildren';
+import { FetchChildrenResponse } from '@/app/utils/types';
 
 export const useChildren = () => {
   const [activePatients, setActivePatients] = useState<number>(0);
@@ -14,11 +14,10 @@ export const useChildren = () => {
       setError(null);
       try {
         console.log('Fetching children from API...');
-        const response: FetchChildrenResponse = await fetchChildren();
+        const response: FetchChildrenResponse = await fetchAllChildren();
         console.log('API response:', response);
 
-        // Ensure that the `child` field is always treated as an array
-        const children: Child[] = response?.child ?? [];  // Default to an empty array if `child` is undefined or null
+        const children: Child [] = response?.child ?? [];
 
         const active = children.filter(child => !child.is_deleted).length;
         const inactive = children.filter(child => child.is_deleted).length;
