+++ kolezza_dash_board/src/app/components/hooks/useGetUsersDashboard.ts
@@ -1,7 +1,7 @@
-import { useState, useEffect } from 'react';
-import { fetchUsers } from '@/app/utils/fetchUser';
+import { fetchUsers } from "@/app/utils/fetchUsers";
+import { useState, useEffect } from "react";
 
-type UserRole = "speech_therapist"|"superadmin";
+type UserRole = "speech_therapist" | "superadmin";
 
 interface User {
   id: string;
@@ -28,26 +28,27 @@ const useGetUsersData = () => {
 
         if (Array.isArray(data)) {
           setUsers(data);
-        } else if (data && typeof data === 'object') {
+        } else if (data && typeof data === "object") {
           const usersArray = data.users || data.results || Object.values(data);
           if (Array.isArray(usersArray)) {
             setUsers(usersArray);
           } else {
-            throw new Error('Unexpected data structure');
+            throw new Error("Unexpected data structure");
           }
         } else {
-          throw new Error('Unexpected data type');
+          throw new Error("Unexpected data type");
         }
       } catch (err) {
-        console.error('Error in useGetUsersData:', err);
-        setError(err instanceof Error ? err.message : 'An error occurred');
+        console.error("Error in useGetUsersData:", err);
+        setError(err instanceof Error ? err.message : "An error occurred");
       } finally {
         setLoading(false);
       }
     };
 
     getUsersData();
-  }, []);
+  }),
+    [];
 
   return { users, loading, error };
 };
