+++ kolezza_dash_board/src/app/dashboard/page.tsx
@@ -2,8 +2,8 @@
 import React, { useState, useEffect } from 'react';
 import { Chart as ChartJS, ArcElement, Tooltip, Legend, BarElement, CategoryScale, LinearScale} from 'chart.js';
 import { Doughnut, Bar } from 'react-chartjs-2';
-import { useFetchTherapists } from '../hooks/useFetchTherapists';
-import { useFetchChildren } from '../hooks/useFetchChildren';
+import { useFetchTherapists } from '../components/hooks/useFetchTherapists';
+import { useFetchChildren } from '../components/hooks/useFetchChildren';
 import Layout from '../Layout';
 import ChartDataLabels from 'chartjs-plugin-datalabels';
 import { ChartOptions} from 'chart.js';
